Python FastApi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-cloud-storage==2.14.0
pillow==10.1.0
python-magic==0.4.27
aiofiles==23.2.1
aiosmtplib==2.0.1
jinja2==3.1.2
python-dateutil==2.8.2